refactor(gateway): extract housekeeping daemons from run.py (slice 30 of #54962)
